+++ src/main/java/com/javampire/openscad/formatter/OpenSCADIndentBuilder.java
@@ -3,6 +3,7 @@ package com.javampire.openscad.formatter;
 import com.intellij.formatting.Indent;
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.codeStyle.CodeStyleSettings;
 import com.intellij.psi.tree.IElementType;
 
 import static com.javampire.openscad.parser.OpenSCADParserDefinition.BRACES_TOKENS;
@@ -11,7 +12,13 @@ import static com.javampire.openscad.psi.OpenSCADTypes.*;
 
 public class OpenSCADIndentBuilder {
 
-    public Indent getChildIndent(ASTNode node) {
+    private final OpenSCADCodeStyleSettings settings;
+
+    public OpenSCADIndentBuilder(final CodeStyleSettings settings) {
+        this.settings = settings.getCustomSettings(OpenSCADCodeStyleSettings.class);
+    }
+
+    public Indent getChildIndent(final ASTNode node) {
         IElementType elementType = node.getElementType();
         ASTNode parent = node.getTreeParent();
 
@@ -24,7 +31,11 @@ public class OpenSCADIndentBuilder {
         if (parentType == BLOCK_OBJ && !BRACES_TOKENS.contains(elementType)) { // Indent all BLOCK_OBJ (surrounded by brackets) children by default
             return Indent.getNormalIndent();
         } else if (parentType == BUILTIN_OBJ && elementType != BUILTIN_OP && elementType != BUILTIN_OBJ_REF) { // Indent statements in BUILTIN_OBJ without brackets
-            return Indent.getNormalIndent();
+            if (settings.INDENT_CASCADING_TRANSFORMATIONS) {
+                return Indent.getNormalIndent();
+            } else {
+                return Indent.getNoneIndent();
+            }
         } else if (parentType == IF_OBJ && (elementType == BUILTIN_OBJ || elementType == FOR_OBJ)) { // Indent statements in IF_OBJ without brackets
             return Indent.getNormalIndent();
         } else if (MATHEMATICAL_EXPR.contains(parentType)) {