+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Main repository is transferred from https://github.com/ncsaba/idea-openscad to h
 - Issue-34 - 2019.05 features not fully supported
 - Issue-38 - 2019.05 list comprehensions does not parse
 - Issue-91 - "include" can be included in a block object
+- Issue-92 - Code style: can disable indent in cascade transformations
 
 ### Changed
 - Dependency updates, cleaning build configuration